pyproject.toml
[project]
name = "telego"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.08"
dependencies = [
    "pyyaml==6.0.2"
]
